# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Salesmen API",
        "description": "Professional REST API for managing salesmen data - PROSIGHT Slovensko Technical Assignment",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/salesmen": {
            "get": {
                "tags": [
                    "Salesmen"
                ],
                "summary": "List all salesmen",
                "description": "Retrieve a paginated list of salesmen with optional sorting",
                "operationId": "getSalesmen",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15,
                            "maximum": 100,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort field and direction (e.g., 'first_name' or '-created_at')",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "first_name"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of salesmen retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SalesmenCollection"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Salesmen"
                ],
                "summary": "Create a new salesman",
                "description": "Create a new salesman record with validation",
                "operationId": "createSalesman",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SalesmanRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Salesman created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SalesmanResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Salesman already exists",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "SalesmanRequest": {
                "required": [
                    "first_name",
                    "last_name",
                    "prosight_id",
                    "email",
                    "gender"
                ],
                "properties": {
                    "first_name": {
                        "description": "First name",
                        "type": "string",
                        "maxLength": 50,
                        "minLength": 2,
                        "example": "Ján"
                    },
                    "last_name": {
                        "description": "Last name",
                        "type": "string",
                        "maxLength": 50,
                        "minLength": 2,
                        "example": "Novák"
                    },
                    "titles_before": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "maxItems": 10,
                        "example": [
                            "Ing.",
                            "Mgr."
                        ]
                    },
                    "titles_after": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "maxItems": 10,
                        "example": [
                            "PhD."
                        ]
                    },
                    "prosight_id": {
                        "description": "5-digit ID",
                        "type": "string",
                        "pattern": "^[0-9]{5}$",
                        "example": "12345"
                    },
                    "email": {
                        "description": "Email",
                        "type": "string",
                        "format": "email",
                        "example": "jan.novak@prosight.sk"
                    },
                    "phone": {
                        "description": "Phone",
                        "type": "string",
                        "example": "+421901234567",
                        "nullable": true
                    },
                    "gender": {
                        "description": "Gender",
                        "type": "string",
                        "enum": [
                            "m",
                            "f"
                        ],
                        "example": "m"
                    },
                    "marital_status": {
                        "type": "string",
                        "enum": [
                            "single",
                            "married",
                            "divorced",
                            "widowed"
                        ],
                        "example": "single",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "SalesmanResponse": {
                "properties": {
                    "data": {
                        "properties": {
                            "id": {
                                "type": "string",
                                "format": "uuid",
                                "example": "01234567-89ab-cdef-0123-456789abcdef"
                            },
                            "first_name": {
                                "type": "string",
                                "example": "Ján"
                            },
                            "last_name": {
                                "type": "string",
                                "example": "Novák"
                            },
                            "display_name": {
                                "type": "string",
                                "example": "Ing. Mgr. Ján Novák PhD."
                            },
                            "titles_before": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                },
                                "example": [
                                    "Ing.",
                                    "Mgr."
                                ]
                            },
                            "titles_after": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                },
                                "example": [
                                    "PhD."
                                ]
                            },
                            "prosight_id": {
                                "type": "string",
                                "example": "12345"
                            },
                            "email": {
                                "type": "string",
                                "example": "jan.novak@prosight.sk"
                            },
                            "phone": {
                                "type": "string",
                                "example": "+421901234567"
                            },
                            "gender": {
                                "type": "string",
                                "example": "m"
                            },
                            "marital_status": {
                                "type": "string",
                                "example": "single"
                            },
                            "created_at": {
                                "type": "string",
                                "format": "date-time"
                            },
                            "updated_at": {
                                "type": "string",
                                "format": "date-time"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "SalesmenCollection": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/SalesmanResponse/properties/data"
                        }
                    },
                    "links": {
                        "properties": {
                            "first": {
                                "type": "string",
                                "nullable": true
                            },
                            "last": {
                                "type": "string",
                                "nullable": true
                            },
                            "prev": {
                                "type": "string",
                                "nullable": true
                            },
                            "next": {
                                "type": "string",
                                "nullable": true
                            }
                        },
                        "type": "object"
                    },
                    "meta": {
                        "properties": {
                            "current_page": {
                                "type": "integer"
                            },
                            "per_page": {
                                "type": "integer"
                            },
                            "total": {
                                "type": "integer"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "properties": {
                    "errors": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "code": {
                                    "type": "string",
                                    "example": "RESOURCE_NOT_FOUND"
                                },
                                "message": {
                                    "type": "string",
                                    "example": "Resource not found"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "first_name": [
                                "The first name field is required."
                            ]
                        }
                    }
                },
                "type": "object"
            }
        }
    },
...
}